Use Int64 literals for backend client-count Refs
Ref(0) creates RefValue{Int32} on 32-bit Julia, which does not convert to
the RefValue{Int64} client-count fields, so register_backend would still
throw a MethodError during __init__ on i686 even with the relaxed kwarg.

Agent-Harness: Devin CLI 3000.10.21
Agent-Model: SWE-2 Max
Agent-Session: local Devin CLI session, no shareable URL (transcript pending write under ~/.local/share/devin/cli/transcripts/)

+++ src/accelerators/Registration.jl
@@ -44,8 +44,8 @@ function register_backend(
             priority,
             pjrt_initialize_function,
             ifrt_initialize_function,
-            Ref(0),
-            Ref(0),
+            Ref(Int64(0)),
+            Ref(Int64(0)),
             preinitialize_setup_function,
         )
         push!(RegisteredBackends, backend)
